Update README.md
Workflow yaml file needed to be updated to a new node version as there is no longer support for 3.4. While this does not affect the current deployment, it is currently not possible to update the website without running into a build error. 

If the Medalign website needs to be updated make sure you first copy the yaml in the readme to .github/worflows/depoly-gthub.yml

This update should also help anyone who wants to publish a new website in the future.

Cheers